Moved eslint formatter
Pulled in latest bindgen branch from core
